created:seperate group info page for group
